Prevent DOM scrolling on kill shortcut
I noticed when implementing clojure-mode on a tall page, whenever I
tried the kill command (Ctrl-k) the kill would happen correctly but my
page would get scrolled down to the bottom. This was because my browser
was focusing on the temporary textarea element created in the
copy-to-clipboard! function. Adding :preventScroll true fixed the
problem for me.